minor: improve handling of routes to be more consistent - refactor DNS (#52)
A collection of minor refactors to organise and simplify how routes are handled in the configuration and the CDK code. This will prompt an update to the from scratch documentation to address these new field names/locations.

Also adds the ability to specify a branch in the config source - this is respected in the pipelines, enabling branch specific config files, another layer of control.